backend/requirements.txt
advertools==0.17.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.1.0
Automat==25.4.16
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.1
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
jq==1.10.0
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.4.0
numpy==2.2.6
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parsel==1.10.0
pillow==11.3.0
Protego==0.5.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
Scrapy==2.13.3
selenium==4.28.1
sentry-sdk==2.34.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.47.2
tk==0.1.0
tldextract==5.3.0
tomli==2.2.1
trio==0.28.0
trio-websocket==0.11.1
Twisted==25.5.0
twython==3.9.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zope.interface==7.2
